[Bugfix]多实例情况下onEvent回调错误 解决:使用LiveCtrl的指针

+++ libp2pimpl/src/main/java/cn/vbyte/p2p/MultiCallbackInterface.java
@@ -12,5 +12,5 @@ public interface MultiCallbackInterface{
      * @param msg 事件参数
      * @param id LiveController的实例ID
      */
-    public void handleMessage(int code, String msg, int id);
+    public void handleMessage(int code, String msg, long id);
 }
